[WIP] Moving adaptivity functionality to a different class called AdaptiveController
